+++ packages/tko.utils.parser/spec/parserBehaviors.js
@@ -35,8 +35,8 @@ function makeBindings (binding, context) {
 
 describe('the bindings parser', function () {
   it('parses bindings with JSON values', function () {
-    var binding_string = 'a: "A", b: 1, b_: -1, c: 2.1, c_:-2.1, d: ["X", "Y"], e: {"R": "V"}, t: true, f: false, n: null',
-      value = new Parser().parse(binding_string);
+    const bindingString = 'a: "A", b: 1, b_: -1, c: 2.1, c_:-2.1, d: ["X", "Y"], e: {"R": "V"}, t: true, f: false, n: null, o'
+    const value = new Parser().parse(bindingString)
     assert.equal(value.a(), 'A', 'string');
     assert.equal(value.b(), 1, 'int');
     assert.equal(value.b_(), -1, '-int');
@@ -49,6 +49,7 @@ describe('the bindings parser', function () {
     assert.equal(value.t(), true, 'true');
     assert.equal(value.f(), false, 'false');
     assert.equal(value.n(), null, 'null');
+    assert.equal(value.o(), undefined, '{o}');
   })
 
   it('works with alphanumerics ä and å (unicode < 0x0100)', function () {
@@ -123,10 +124,17 @@ describe('the bindings parser', function () {
   })
 
   it('parses object: attr: {name: value}', function () {
-    var binding = 'attr: { klass: kValue }',
-      context = ctxStub({ kValue: 'Sam' }),
-      bindings = new Parser().parse(binding, context);
+    const binding = 'attr: { klass: kValue }'
+    const context = ctxStub({ kValue: 'Sam' })
+    const bindings = new Parser().parse(binding, context)
     assert.equal(bindings.attr().klass, 'Sam')
+  })
+
+  it('parses object: attr: {name}', function () {
+    const binding = 'attr: { kv }'
+    const context = ctxStub({ kv: 'Sam' })
+    const bindings = new Parser().parse(binding, context)
+    assert.equal(bindings.attr().kv, 'Sam')
   })
 
   it('parses object: attr: {name: observable(value)}', function () {